Vocab buttons now also have blacke border
